resolve classes longas e sem contexto
