+++ src/br/ufal/ic/p2/wepayu/ModuleSindicato/Service/SindicatoService.java
@@ -1,10 +1,14 @@
 package br.ufal.ic.p2.wepayu.ModuleSindicato.Service;
+import br.ufal.ic.p2.wepayu.Core.Exceptions.EmpregadoNaoEhSindicalizado;
+import br.ufal.ic.p2.wepayu.Core.Exceptions.ValorInvalido;
+import br.ufal.ic.p2.wepayu.ModuleEmpregado.model.Empregado;
 import br.ufal.ic.p2.wepayu.ModuleSindicato.Classes.MembroSindicato;
 import java.text.DecimalFormat;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 
+import static br.ufal.ic.p2.wepayu.ModuleEmpregado.Service.EmpregadoValidations.validaSindicalizacao;
 import static br.ufal.ic.p2.wepayu.ModuleSindicato.Service.SindicatoValidations.*;
 import static br.ufal.ic.p2.wepayu.Sistema.*;
 
@@ -41,6 +45,53 @@ public class SindicatoService {
         }
         DecimalFormat df = new DecimalFormat("#,##0.00");
         return df.format(taxasDeServicoTotais).replace('.', ',');
+    }
+
+    public static void alteraSindicalizado(Empregado empregado, String valor) throws ValorInvalido {
+        if (valor.equalsIgnoreCase("true")) {
+            empregado.SetisSindicalizado(true);
+        } else if (valor.equalsIgnoreCase("false")) {
+            empregado.SetisSindicalizado(false);
+        } else {
+            throw new ValorInvalido("Valor inválido");
+        }
+    }
+
+    public static void AlteraEmpregadoSindicato(String id, String atributo, String valor) throws Exception {
+        Empregado empregado = listaEmpregados.get(id);
+        if (valor.equalsIgnoreCase("false"))
+            empregado.SetisSindicalizado(false);
+
+        if (valor.equalsIgnoreCase("true"))
+            empregado.SetisSindicalizado(true);
+
+        if (!empregado.getTipo().equalsIgnoreCase("comissionado"))
+            throw new EmpregadoNaoEhSindicalizado("Empregado nao eh sindicalizado.");
+    }
+
+    public static void SindicalizaEmpregado(String id, String atributo, String valor, String idSindical, String taxaSindical) throws Exception {
+        validaSindicalizacao(id,atributo,valor,idSindical,taxaSindical);
+        alteraEmpregadoValidation(idSindical);
+        Empregado empregado = listaEmpregados.get(id);
+        empregado.SetisSindicalizado(true);
+        criaVinculoSindical(id,idSindical,taxaSindical);
+        listaEmpregados.put(id,empregado);
+    }
+
+    public static String obterIdSindicato(Empregado empregado) throws EmpregadoNaoEhSindicalizado {
+        if (!empregado.getIsSindicalizado()) {
+            throw new EmpregadoNaoEhSindicalizado("Empregado nao eh sindicalizado.");
+        }
+        MembroSindicato sindicalista = listaDeSindicalizados.get(empregado.getId());
+        return sindicalista.getIdSindical();
+    }
+
+    public static String obterTaxaSindical(Empregado empregado) throws EmpregadoNaoEhSindicalizado {
+        if (!empregado.getIsSindicalizado()) {
+            throw new EmpregadoNaoEhSindicalizado("Empregado nao eh sindicalizado.");
+        }
+        MembroSindicato sindicalista = listaDeSindicalizados.get(empregado.getId());
+        return sindicalista.getTaxaSindical();
     }
 
 }
